+++ app/src/main/java/com/mine/autoshine/MainActivity.java
@@ -27,6 +27,8 @@ import android.widget.Switch;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import com.mine.autoshine.services.ShineService;
+
 public class MainActivity extends Activity {
 
     private static final String TAG = "MainActivity";
